+++ pkg/roles/api/middleware/middleware_log.go
@@ -1,4 +1,4 @@
-package api
+package middleware
 
 import (
 	"bufio"
