+++ pkg/roles/api/middleware/middleware_recover.go
@@ -1,4 +1,4 @@
-package api
+package middleware
 
 import (
 	"net/http"
